Disable the 'check' target when included as a sub-project
This disables adding the 'check' target when ngtcp2 is included as a
subproject (via `add_directory(ngtcp2)` from another project) so that
the parent project can define its own `check` target.

Without this patch attempting to define a `check` target in the parent
will cause a cmake error because of duplicate targets.